+++ kmath-ast/src/commonMain/kotlin/space/kscience/kmath/ast/parser.kt
@@ -18,10 +18,10 @@ import com.github.h0tk3y.betterParse.parser.ParseResult
 import com.github.h0tk3y.betterParse.parser.Parser
 import space.kscience.kmath.expressions.MST
 import space.kscience.kmath.expressions.Symbol
-import space.kscience.kmath.operations.FieldOperations
-import space.kscience.kmath.operations.GroupOperations
+import space.kscience.kmath.operations.FieldOps
+import space.kscience.kmath.operations.GroupOps
 import space.kscience.kmath.operations.PowerOperations
-import space.kscience.kmath.operations.RingOperations
+import space.kscience.kmath.operations.RingOps
 
 /**
  * better-parse implementation of grammar defined in the ArithmeticsEvaluator.g4.
@@ -60,7 +60,7 @@ public object ArithmeticsEvaluator : Grammar<MST>() {
         .or(binaryFunction)
         .or(unaryFunction)
         .or(singular)
-        .or(-minus and parser(ArithmeticsEvaluator::term) map { MST.Unary(GroupOperations.MINUS_OPERATION, it) })
+        .or(-minus and parser(ArithmeticsEvaluator::term) map { MST.Unary(GroupOps.MINUS_OPERATION, it) })
         .or(-lpar and parser(ArithmeticsEvaluator::subSumChain) and -rpar)
 
     private val powChain: Parser<MST> by leftAssociative(term = term, operator = pow) { a, _, b ->
@@ -72,9 +72,9 @@ public object ArithmeticsEvaluator : Grammar<MST>() {
         operator = div or mul use TokenMatch::type
     ) { a, op, b ->
         if (op == div)
-            MST.Binary(FieldOperations.DIV_OPERATION, a, b)
+            MST.Binary(FieldOps.DIV_OPERATION, a, b)
         else
-            MST.Binary(RingOperations.TIMES_OPERATION, a, b)
+            MST.Binary(RingOps.TIMES_OPERATION, a, b)
     }
 
     private val subSumChain: Parser<MST> by leftAssociative(
@@ -82,9 +82,9 @@ public object ArithmeticsEvaluator : Grammar<MST>() {
         operator = plus or minus use TokenMatch::type
     ) { a, op, b ->
         if (op == plus)
-            MST.Binary(GroupOperations.PLUS_OPERATION, a, b)
+            MST.Binary(GroupOps.PLUS_OPERATION, a, b)
         else
-            MST.Binary(GroupOperations.MINUS_OPERATION, a, b)
+            MST.Binary(GroupOps.MINUS_OPERATION, a, b)
     }
 
     override val rootParser: Parser<MST> by subSumChain