+++ kmath-histograms/src/commonTest/kotlin/space/kscience/kmath/histogram/MultivariateHistogramTest.kt
@@ -5,6 +5,7 @@
 
 package space.kscience.kmath.histogram
 
+import space.kscience.kmath.nd.DefaultStrides
 import space.kscience.kmath.operations.invoke
 import space.kscience.kmath.real.DoubleVector
 import kotlin.random.Random
@@ -69,7 +70,7 @@ internal class MultivariateHistogramTest {
             }
             val res = histogram1 - histogram2
             assertTrue {
-                strides.indices().all { index ->
+                DefaultStrides(shape).indices().all { index ->
                     res.values[index] <= histogram1.values[index]
                 }
             }